Improve the pending reason description for various QOS limits.

+++ src/slurmctld/acct_policy.h
@@ -91,7 +91,7 @@ extern bool acct_policy_validate(job_desc_msg_t *job_desc,
 				 struct part_record *part_ptr,
 				 slurmdb_association_rec_t *assoc_in,
 				 slurmdb_qos_rec_t *qos_ptr,
-				 uint16_t *state_reason,
+				 uint32_t *state_reason,
 				 acct_policy_limit_set_t *acct_policy_limit_set,
 				 bool update_call);
